Split into a library
